% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptest.R
\name{adaptest}
\alias{adaptest}
\title{adaptest R package: Data-adaptive statistics for high-dimensional multiple testing}
\usage{
adaptest()
}
\description{
Data-adaptive test statistics represent a general methodology for
performing multiple hypothesis testing on effects sizes while maintaining
honest statistical inference when operating in high-dimensional settings.
The utilities provided here extend the use of this general methodology to
many common data analytic challenges that arise in modern computational and
genomic biology.
}
\references{
\describe{
  \item{Bioconductor:}{\url{https://bioconductor.org/packages/adaptest}}
  \item{Documentation:}{\url{https://wilsoncai1992.github.io/adaptest/}}
  \item{GitHub repository:}{\url{https://wilsoncai1992.github.io/adaptest/}}
}
}
