% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotmle.R
\name{biotmle}
\alias{biotmle}
\title{biotmle}
\usage{
biotmle()
}
\description{
Package title
}
\references{
\describe{
  \item{Documentation:}{https://code.nimahejazi.org/biotmle}
  \item{GitHub repository:}{https://github.com/nhejazi/biotmle}
  \item{Bioconductor:}{https://bioconductor.org/packages/biotmle}
}
}
