% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hal9001.R
\name{hal9001}
\alias{hal9001}
\title{`hal9001` R package}
\usage{
hal9001()
}
\description{
Scalable and Efficient Highly Adaptive Lasso
}
\references{
\describe{
  \item{Documentation:}{\url{https://jeremyrcoyle.github.io/hal9001}}
  \item{GitHub repository:}{\url{https://github.com/njeremyrcoyle/hal9001}}
}
}
