% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haldensify.R
\name{haldensify}
\alias{haldensify}
\title{haldensify R package: Conditional density estimation with the Highly Adaptive Lasso}
\usage{
haldensify()
}
\description{
Conditional density estimation is a longstanding and challenging problem in
statistical theory, and numerous proposals exist for optimally estimating
such complex functions. An algorithm for nonparametric estimation of
conditional densities based on a pooled hazard regression formulation is
implemented based on the highly adaptive lasso (HAL), an optimal algorithm
for efficient estimation with guarantees of fast convergence rates.
}
\references{
\describe{
  \item{Documentation:}{\url{https://code.nimahejazi.org/haldensify}}
  \item{GitHub repository:}{\url{https://github.com/nhejazi/haldensify}}
}
}
