% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyvim.R
\name{methyvim}
\alias{methyvim}
\title{`methyvim` R package}
\usage{
methyvim()
}
\description{
Targeted Variable Importance for Differential Methylation Analysis
}
\references{
\describe{
  \item{Bioconductor:}{\url{https://bioconductor.org/packages/methyvim}}
  \item{Documentation:}{\url{https://code.nimahejazi.org/methyvim}}
  \item{GitHub repository:}{\url{https://github.com/nhejazi/methyvim}}
}
}
