% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nima.R
\name{nima}
\alias{nima}
\title{`nima` R package}
\usage{
nima()
}
\description{
Nima Hejazi's R Toolbox
}
\references{
\describe{
  \item{CRAN:}{\url{https://CRAN.R-project.org/package=nima}}
  \item{Documentation:}{\url{https://code.nimahejazi.org/nima}}
  \item{GitHub repository:}{\url{https://github.com/nhejazi/nima}}
}
}
