% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl3.R
\name{sl3}
\alias{sl3}
\title{`sl3` R package}
\usage{
sl3()
}
\description{
Pipelines for Machine Learning and Super Learning
}
\references{
\describe{
  \item{Documentation:}{\url{https://sl3.tlverse.org}}
  \item{GitHub repository:}{\url{https://github.com/tlverse/sl3}}
}
}
