% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survtmle.R
\name{survtmle}
\alias{survtmle}
\title{survtmle R package: Compute Targeted Minimum Loss-Based Estimates in Right-Censored Survival Settings}
\usage{
survtmle()
}
\description{
Targeted estimates of marginal cumulative incidence in survival settings
with and without competing risks, including estimators that respect bounds
(Benkeser, Carone, and Gilbert. Statistics in Medicine, 2017.
<doi:10.1002/sim.7337>).
}
\references{
\describe{
  \item{CRAN:}{\url{https://CRAN.R-project.org/package=survtmle}}
  \item{Documentation:}{\url{https://benkeser.github.io/survtmle}}
  \item{GitHub repository:}{\url{https://github.com/benkeser/survtmle}}
}
}
