% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txshift.R
\name{txshift}
\alias{txshift}
\title{`txshift` R package}
\usage{
txshift()
}
\description{
Targeted Minimum Loss-Based Estimation of the Causal Effect of Stochastic
Interventions and Variable Importance Analysis
}
\references{
\describe{
  \item{Documentation:}{\url{https://code.nimahejazi.org/txshift}}
  \item{GitHub repository:}{\url{https://github.com/nhejazi/txshift}}
}
}
