% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biotmle.R
\name{biotmle}
\alias{biotmle}
\title{`biotmle` R package}
\usage{
biotmle()
}
\description{
Moderated and Targeted Statistical Learning for Biomarker Discovery
}
\references{
\describe{
  \item{Bioconductor:}{\url{https://bioconductor.org/packages/biotmle}}
  \item{Documentation:}{\url{https://code.nimahejazi.org/biotmle}}
  \item{GitHub repository:}{\url{https://github.com/nhejazi/biotmle}}
}
}
