% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origami.R
\name{origami}
\alias{origami}
\title{`origami` R package}
\usage{
origami()
}
\description{
Generalized Framework for Cross-Validation
}
\references{
\describe{
  \item{CRAN:}{\url{https://CRAN.R-project.org/package=origami}}
  \item{Documentation:}{\url{https://origami.tlverse.org}}
  \item{GitHub repository:}{\url{https://github.com/tlverse/origami}}
}
}
