% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origami.R
\name{origami}
\alias{origami}
\title{origami R package: Generalized Framework for Cross-Validation}
\usage{
origami()
}
\description{
Provides a general framework for the application of cross-validation schemes
to particular functions. By allowing arbitrary lists of results, origami
accommodates a range of cross-validation applications.
}
\references{
\describe{
  \item{CRAN:}{\url{https://CRAN.R-project.org/package=origami}}
  \item{Documentation:}{\url{https://origami.tlverse.org}}
  \item{GitHub repository:}{\url{https://github.com/tlverse/origami}}
}
}
