% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl3.R
\name{sl3}
\alias{sl3}
\title{sl3 R package: Pipelines for machine learning and Super Learning}
\usage{
sl3()
}
\description{
Modern implementation of the Super Learner prediction algorithm, alongside
a framework for general-purpose machine learning with pipelines.
}
\references{
\describe{
  \item{Documentation:}{\url{https://tlverse.org/sl3}}
  \item{GitHub repository:}{\url{https://github.com/tlverse/sl3}}
}
}
