% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survtmle.R
\name{survtmle}
\alias{survtmle}
\title{`survtmle` R package}
\usage{
survtmle()
}
\description{
Compute Targeted Minimum Loss-Based Estimates in Right-Censored Survival
Settings
}
\references{
\describe{
  \item{CRAN:}{\url{https://CRAN.R-project.org/package=survtmle}}
  \item{Documentation:}{\url{https://benkeser.github.io/survtmle}}
  \item{GitHub repository:}{\url{https://github.com/benkeser/survtmle}}
}
}
