% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txshift.R
\name{txshift}
\alias{txshift}
\title{txshift R package: Targeted Minimum Loss-Based Estimation of the Causal Effect of Stochastic Interventions and Variable Importance Analysis}
\usage{
txshift()
}
\description{
Facilities for computing targeted maximum likelihood estimates of
population-level causal effects of continuous-valued interventions defined by
stochastic assignment mechanisms. To handle applications with censored data
(i.e., two-stage designs), an inverse probability of censoring weighted
(IPCW) targeted maximum likelihood estimator for the stochastic treatment
shift parameter is implemented as well.
}
\references{
\describe{
  \item{Documentation:}{\url{https://code.nimahejazi.org/txshift}}
  \item{GitHub repository:}{\url{https://github.com/nhejazi/txshift}}
}
}
